refactor#67 PutRetrofitWork 코루틴 적용
