+++ app/src/main/java/com/studymate154/studymate/MyPage/MyPostActivity.kt
@@ -9,7 +9,7 @@ import android.util.Log
 import androidx.recyclerview.widget.DividerItemDecoration
 import androidx.recyclerview.widget.LinearLayoutManager
 import com.studymate154.studymate.board.BoardInsideActivity
-import com.studymate154.studymate.board.BoardListAdapter
+import com.studymate154.studymate.board.BoardAdapter.BoardListAdapter
 import com.studymate154.studymate.Model.GetBoardModel
 import com.studymate154.studymate.board.PostRetrofitAPI
 import com.studymate154.studymate.databinding.ActivityMyPostBinding
