+++ app/src/main/java/com/studymate154/studymate/board/BoardWriteActivity.kt
@@ -92,7 +92,7 @@ class BoardWriteActivity : AppCompatActivity() {
 
         }
 
-        // @post
+        // 게시글 post하기
         binding.postBtn.setOnClickListener {
             boardModel.title = binding.editTitle.text.toString()
             boardModel.content = binding.editContent.text.toString()
