+++ app/src/main/java/com/studymate154/studymate/board/RetrofitWork/BoardPostRetrofitWork.kt
@@ -2,12 +2,9 @@ package com.studymate154.studymate.board
 
 import android.util.Log
 import com.studymate154.studymate.Model.BoardWriteModel
-import com.studymate154.studymate.signUp.SignUpResponseBody
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.launch
-import retrofit2.Call
-import retrofit2.Response
 
 class StudyRetrofitWork(private val userToken: String, private val postInfo: BoardWriteModel) {
     fun work() {
@@ -23,7 +20,7 @@ class StudyRetrofitWork(private val userToken: String, private val postInfo: Boa
             val response = service.addPostByEnqueue("Bearer $userToken", postInfo)
 
             if(response.isSuccessful){
-                Log.d("Login","로그인 통신 성공")
+                Log.d("Login","게시글 작성 성공")
             }
 
         }
