+++ app/src/main/res/layout/activity_board_inside.xml
@@ -21,10 +21,6 @@
               android:layout_marginStart="10dp"
               android:src="@drawable/baseline_arrow_back_24" />
 
-          <View
-              android:layout_width="0dp"
-              android:layout_height="0dp"
-              android:layout_weight="1"/>
 
           <TextView
               android:id="@+id/category"
@@ -32,9 +28,14 @@
               android:layout_height="wrap_content"
               android:text="자유게시판"
               android:layout_gravity="center"
-              android:layout_marginEnd="30dp"
+              android:layout_marginStart="20dp"
               android:textSize="17sp"
               android:textColor="#915FC1"/>
+
+          <View
+              android:layout_width="0dp"
+              android:layout_height="0dp"
+              android:layout_weight="1"/>
 
           <ImageView
               android:id="@+id/menuImg"
@@ -59,6 +60,7 @@
     <LinearLayout
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
+        android:layout_marginTop="10dp"
         android:orientation="horizontal">
 
         <de.hdodenhof.circleimageview.CircleImageView
